Some fixes for OSX

+++ autogen.sh
@@ -1,11 +1,25 @@
 #!/bin/bash
 
-LIBTOOL=libtool
-LIBTOOLIZE=libtoolize
+if [ -z $LIBTOOLIZE ] ; then
+    LIBTOOLIZE=$(type -P libtoolize)
+    if [ -z $LIBTOOLIZE ] ; then
+        LIBTOOLIZE=$(type -P glibtoolize)
+    fi
+fi
+if [ -z $LIBTOOL ] ; then
+    LIBTOOL=$(type -P "${LIBTOOLIZE%???}")
+fi
+
+if [ -z $LIBTOOL ] || [ -z $LIBTOOLIZE ] ; then
+    echo "Unable to find working libtool. [$LIBTOOL][$LIBTOOLIZE]"
+    exit 1
+fi
+
 
 # Delete the old libtool output
 rm -rf libltdl src/sst/core/libltdl
 
+echo "Running ${LIBTOOLIZE}..."
 $LIBTOOLIZE --automake --copy --ltdl
 
 if test -d libltdl; then
